adding save and load
had to break up these two commits because it was tough considering whether the process would be interactive as far as picking the filename, or if save would require a filename. rather than be concerned about whether we'd somehow accidentally overwrite the wrong config this seems like a good compromise. require a path for both so we can blame the user instead :)