configuration timing
there was an issue where when adding a new site, I'd still have to restart appserve to initiate the new configuration post-adding the new site. this should alter how the sites are being loaded so that you no longer require restart to instantiate a new site. i hope. idk. it's saturday.